Update validation for friend list to handle list editing
