+++ src/validation/friendList.js
@@ -10,12 +10,12 @@ export function validateFriendListInput(data) {
     let errors = {};
 
     // Title check
-    if (isEmpty(data.title)) {
+    if (data.title && isEmpty(data.title)) {
         errors.title = "List title is required";
     }
 
     // Friends array check
-    if (isEmpty(data.friends)) {
+    if (data.friends && isEmpty(data.friends)) {
         errors.friends = "List should have at least one friend";
     }
